+++ src/main/java/com/utn/space/venueaapi/repository/CancellationPolicyRepository.java
@@ -3,6 +3,6 @@ package com.utn.space.venueaapi.repository;
 import com.utn.space.venueaapi.model.CancellationPolicy;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CancellationPolicyRepository extends JpaRepository<CancellationPolicy,Long> {
+public interface CancellationPolicyRepository extends JpaRepository<CancellationPolicy,Integer> {
 
 }
